pyproject.toml
[tool.poetry]
name = "function-calling"
version = "0.1.0"
description = ""
authors = ["SijiaCui <cuisijia0908@gmail.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.52.2"
google-generativeai = "^0.8.3"
anthropic = "^0.37.1"
mysql-connector-python = "^9.1.0"
yfinance = "^0.2.48"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
